fix BreadCrump navigation
